Move GrpcServerProcess to shared module; sort imports consistently
